fix: move middleware
